+++ pytest_reportportal/config.py
@@ -1,4 +1,5 @@
 """This module contains class that stores RP agent configuration data."""
+import warnings
 
 from distutils.util import strtobool
 from os import getenv
@@ -69,14 +70,55 @@ class AgentConfig(object):
                                            'rp_project')
         self.rp_rerun_of = self.find_option(pytest_config,
                                             'rp_rerun_of')
-        self.rp_retries = int(self.find_option(pytest_config,
-                                               'retries'))
         self.rp_skip_connection_test = str(
             self.find_option(pytest_config,
                              'rp_skip_connection_test')).lower() in (
                                            'true', '1', 'yes', 'y')
-        self.rp_uuid = getenv('RP_UUID') or self.find_option(pytest_config,
-                                                             'rp_uuid')
+
+        rp_api_retries_str = self.find_option(pytest_config, 'rp_api_retries')
+        rp_api_retries = rp_api_retries_str and int(rp_api_retries_str)
+        if rp_api_retries and rp_api_retries > 0:
+            self.rp_api_retries = rp_api_retries
+        else:
+            rp_api_retries_str = self.find_option(pytest_config, 'retries')
+            rp_api_retries = rp_api_retries_str and int(rp_api_retries_str)
+            if rp_api_retries and rp_api_retries > 0:
+                self.rp_api_retries = rp_api_retries
+                warnings.warn(
+                    'Parameter `retries` is deprecated since 5.1.9 '
+                    'and will be subject for removing in the next '
+                    'major version. Use `rp_api_retries` argument '
+                    'instead.',
+                    DeprecationWarning,
+                    2
+                )
+            else:
+                self.rp_api_retries = 0
+
+        self.rp_api_key = getenv(
+            'RP_API_KEY') or self.find_option(pytest_config, 'rp_api_key')
+        if not self.rp_api_key:
+            self.rp_api_key = getenv(
+                'RP_UUID') or self.find_option(pytest_config, 'rp_uuid')
+            if self.rp_api_key:
+                warnings.warn(
+                    'Parameter `rp_uuid` is deprecated since 5.1.9 '
+                    'and will be subject for removing in the next '
+                    'major version. Use `rp_api_key` argument '
+                    'instead.',
+                    DeprecationWarning,
+                    2
+                )
+            else:
+                warnings.warn(
+                    'Argument `rp_api_key` is `None` or empty string, '
+                    'that is not supposed to happen because Report '
+                    'Portal is usually requires an authorization key. '
+                    'Please check your configuration.',
+                    RuntimeWarning,
+                    2
+                )
+
         rp_verify_ssl = self.find_option(pytest_config, 'rp_verify_ssl', True)
         try:
             self.rp_verify_ssl = bool(strtobool(rp_verify_ssl))