+++ tests/helpers/utils.py
@@ -24,7 +24,7 @@ DEFAULT_VARIABLES = {
     'rp_launch': 'Pytest',
     'rp_endpoint': 'http://localhost:8080',
     'rp_project': 'default_personal',
-    'rp_uuid': 'test_uuid',
+    'rp_api_key': 'test_api_key',
     'rp_skip_connection_test': 'True'
 }