+++ tests/unit/conftest.py
@@ -60,7 +60,7 @@ def mocked_config():
     mocked_config.option = mock.create_autospec(Config)
     mocked_config.option.rp_project = 'default_personal'
     mocked_config.option.rp_endpoint = 'http://docker.local:8080/'
-    mocked_config.option.rp_uuid = mock.sentinel.rp_uuid
+    mocked_config.option.rp_api_key = mock.sentinel.rp_api_key
     mocked_config.option.rp_log_batch_size = -1
     mocked_config.option.retries = -1
     mocked_config.option.rp_hierarchy_dirs_level = '0'